% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyData.R
\name{tidyDataMerge}
\alias{tidyDataMerge}
\title{Merging tidy data.frames from same domain and indicator
Merge list obtained from a run of 'tidyData' into a single data frame
including new columns for country, domain and indicator}
\usage{
tidyDataMerge(
  tidy.list,
  domain,
  indicator,
  save = FALSE,
  name.out = "output_merged"
)
}
\arguments{
\item{tidy.list}{list from 'tidyData' function (run on multiple sheets)}

\item{domain}{string of domain e.g. "work", "health" etc.}

\item{indicator}{string of indicator e.g. "employment", "median earnings" etc.}

\item{save}{excel spreadsheet output required (TRUE / FALSE)}

\item{name.out}{name of excel spreadsheet to be ouput (requires save = TRUE)}
}
\value{
data.frame of merged list, and (optional) excel spreadsheet
}
\description{
Merging tidy data.frames from same domain and indicator
Merge list obtained from a run of 'tidyData' into a single data frame
including new columns for country, domain and indicator
}
