% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyData.R
\name{tidyDataMergeAll}
\alias{tidyDataMergeAll}
\title{Merging a list of tidy data frames from different domains and indicators
Merge data frames coersed into list, obtained from multiple runs of
'tidyDataMerge' into a single data frame including new column for 'measure' (i.e.
percentage, median income etc.)}
\usage{
tidyDataMergeAll(tidy.list, save = FALSE, name.out = "output_merged")
}
\arguments{
\item{tidy.list}{list of data frames obtained from 'tidyDataMerge' function (run
on different domains and indicator)}

\item{save}{excel spreadsheet output required (TRUE / FALSE)}

\item{name.out}{name of excel spreadsheet to be ouput (requires save = TRUE)}
}
\value{
data.frame of merged list, and (optional) excel spreadsheet
}
\description{
Merging a list of tidy data frames from different domains and indicators
Merge data frames coersed into list, obtained from multiple runs of
'tidyDataMerge' into a single data frame including new column for 'measure' (i.e.
percentage, median income etc.)
}
